requirements.txt
selenium==4.15.0
allure-pytest==2.13.2
pytest==7.4.3
webdriver-manager==4.0.1
requests==2.31.0
beautifulsoup4==4.12.2